### 2016-10-08 #89 @whedon
[PRE REVIEW]: edarf: Exploratory Data Analysis using Random Forests
**Submitting author:** @zmjones (<a href="http://orcid.org/0000-0002-7523-0471">Zachary Jones</a>)
**Repository:** <a href="http://github.com/zmjones/edarf/" target ="_blank">http://github.com/zmjones/edarf/</a>
**Version:** 1.0.0
