### 2016-10-11 #89 @zmjones
i asked on twitter. most people i would suggest know me personally and probably wouldn't be really objective.

that said, anybody in the https://github.com/mlr-org group would be capable of reviewing it (i work with them though). i think also anybody in https://github.com/ropensci.

similar packages include caret, ggrandomforests, mlr. those package authors might be willing.
