### 2016-10-12 #89 @cMadan
@zmjones - thanks for the suggestions!

@ehrlinger, @PhilippPro, @ja-thomas - would you be interested in being a reviewer for this package? (see http://joss.theoj.org/about for the reviewer guidelines)
