### 2016-10-12 #89 @PhilippPro
I can do this, I am generally also interested in random forests.

At first glance it seems like most functionalities are already given in the packages, of course not with a unified interface. And of course they could be extended.
Moreover the ranger package is missing and party is not exact random forest. And there are much more packages that are similar to random forest (like party), see here for a long list in [my blog](http://philipppro.github.io/More_complete_list/).
