### 2016-10-12 #89 @zmjones
hey @PhilippPro. thanks!

do check out their [reviewing criteria](http://joss.theoj.org/about).

it appears that i didn't do a very good job of putting in a "need" bit in the paper, and although i have a vignette i don't have it linked in the readme. do you know a way to do that btw? i generated it locally and then pushed that directory, but i don't know how to get github to render it without creating a new gh-pages branch.

as far as novelty is concerned (though again, see the reqs, as this isn't needed). local importance (and hence class specific importance) is only available for `randomForest` (and isn't as general), and partial dependence isn't really available in any of them (and not in this general form). i also have a bunch of visualization functionality in there that doesn't exist elsewhere.

except for mlr of course. i wrote this package before i started working on mlr, and although this is nearly a strict subset of the functionality in mlr, it does get used quite a bit by people in my field because the start-up costs are really low. e.g., load your data as normal, one line for the random forest call (most people don't tune really, but at least with random forests this isn't as big of an issue), and then immediately on to interpretation stuff with edarf.
