### 2016-10-12 #89 @PhilippPro
What do you mean with "need" bit?
Sorry, I think I cannot help with the rendering in github. Maybe ask @jakob-r.

I will try to do the review until the end of next week.
