### 2016-10-12 #89 @zmjones
Under documentation it says
"
A statement of need

The authors should clearly state what problems the software is designed to solve and who the target audience is.
"
so this is one of the reviewing criteria.

i will sort out the vignette issue today, and submit to cran as well.

i'll look into adding support for ranger too. Arborist looks really nice. I hadn't seen that before. My reading of the reviewing criteria is that this is not an issue, but you can decide for yourself of course.
