### 2016-10-12 30499d3
Update paper.md
