### 2016-10-13 #92 @cMadan
@PhilippPro - The reviewer guidelines are here: http://joss.theoj.org/about#reviewer_guidelines

For the most part you need to see if the statements above in the first post are true and check things off accordingly (as you've already started). If something isn't the case, or you otherwise have a suggestion, just post a comment to @zmjones here (or ask me) and we'll sort things out until the submission meets all of JOSS' guidelines.
