### 2016-10-13 #92 @PhilippPro
I have posted in his issue tracker: https://github.com/zmjones/edarf/issues

I do not understand the last point, which also is not a complete sentence:

> References: Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?
