### 2016-10-13 #92 @cMadan
@PhilippPro - perfect, thank you!

As for the references, I think there's a comma missing, but it otherwise reads fine to me. For each of the references in the compiled paper (https://github.com/openjournals/joss-reviews/files/525302/10.21105.joss.00092.pdf), is there a DOI. Here there are two references, neither of which have a DOI, but I expect both of them should have DOIs, so @zmjones will need to fix that in the bib file. (E.g., older papers may simply not have a DOI to list.)
