### 2016-10-13 7d73a88
add test for extract_proximity
