### 2016-10-13 0c34377
extract_prox readme typo
