### 2016-10-13 311d171
remove dontrun from plot_prox
