+++ pkg/man/plot_prox.Rd
@@ -48,7 +48,6 @@ a ggplot object
 Plot principle components of the proximity matrix
 }
 \examples{
-\dontrun{
 fit <- randomForest(hp ~ ., mtcars, proximity = TRUE)
 prox <- extract_proximity(fit)
 pca <- prcomp(prox, scale = TRUE)
@@ -58,8 +57,6 @@ fit <- randomForest(Species ~ ., iris, proximity = TRUE)
 prox <- extract_proximity(fit)
 pca <- prcomp(prox, scale = TRUE)
 plot_prox(pca, color = iris$Species, color_label = "Species", size = 2)
-}
-
 }
 \references{
 https://github.com/vqv/ggbiplot
