### 2016-10-13 adef7cc
add extract_proximity example
