### 2016-10-13 6dc8a35
add seealso links to plot functions
