+++ .travis.yml
@@ -1,6 +1,13 @@
 language: r
 sudo: required
+cache: packages
 
-r_github_packages:
- - zmjones/party/pkg
- - swager/randomForestCI
+install:
+  - Rscript -e 'install.packages(c("devtools","roxygen2","testthat")); devtools::install_deps("pkg")'
+script:
+  - Rscript -e 'devtools::check("pkg")'
+
+notifications:
+  email:
+    on_success: change
+    on_failure: change
