### 2016-10-13 7a291b0
fix figure location in paper
