### 2016-10-13 450e0f5
add knitr to travis install list
