### 2016-10-13 5936e19
update readme and re-add paper pdf
