+++ .travis.yml
@@ -3,7 +3,7 @@ sudo: required
 cache: packages
 
 install:
-  - Rscript -e 'install.packages(c("devtools", "roxygen2", "testthat", "knitr")); devtools::install_deps("pkg")'
+  - Rscript -e 'install.packages(c("devtools", "roxygen2", "testthat", "knitr", "rmarkdown")); devtools::install_deps("pkg")'
 script:
   - Rscript -e 'devtools::check("pkg")'
 
