### 2016-10-13 30a9c68
add 'need' and 'use' to paper intro
