+++ paper/paper.bib
@@ -1,10 +1,11 @@
 @article{friedman2001greedy,
-  title={Greedy function approximation: a gradient boosting machine},
-  author={Friedman, Jerome H},
-  journal={Annals of statistics},
-  pages={1189--1232},
-  year={2001},
-  publisher={JSTOR}
+  title =	 {Greedy function approximation: a gradient boosting
+                  machine},
+  author =	 {Friedman, Jerome H},
+  journal =	 {Annals of statistics},
+  pages =	 {1189--1232},
+  year =	 2001,
+  publisher =	 {JSTOR}
 }
 
 @article{gerber2014can,
@@ -17,3 +18,15 @@
   publisher =	 {Wiley Online Library}
 }
 
+@inproceedings{jones2015exploratory,
+  title =	 {Exploratory data analysis using random forests},
+  author =	 {Jones, Zachary and Linder, Fridolin},
+  booktitle =	 {Prepared for the 73rd annual MPSA conference},
+  year =	 2015
+}
+
+@article{jones2016there,
+  title =	 {Is There More Violence in the Middle?},
+  author =	 {Jones, Zachary M and Lupu, Yonatan},
+  year =	 2016
+}
