+++ paper/paper.md
@@ -20,6 +20,8 @@ bibliography: paper.bib
 
 This package contains functions useful for exploratory data analysis using random forests, which can be fit using the [randomForest](https://cran.r-project.org/package=randomForest), [randomForestSRC](https://cran.r-project.org/package=randomForestSRC), or [party](https://cran.r-project.org/package=party) packages. These functions can compute the partial dependence of covariates (individually or in combination) on the fitted forests' predictions, the permutation importance of covariates, as well as the distance between data points according to the fitted model.
 
+Random forests are an attractive method for social scientists. Random forests have only a few important tuning parameters and can be adapted to do classification, regression, and clustering. Many research tasks require interpretable models, and, although methods for interpreting random forests exist, in the most popular packages for fitting random forests these methods are available inconsistently. For example although there are methods for computing permutation importance in the `randomForest`, `randomForestSRC` and `party` packages, only `randomForest` can compute local importance. None of the packages can compute permutation importance for groups of covariates. Similarly partial dependence is only implemented in `randomForest`, and it has limited functionality compared to the functions provided herein. This software has been used in @jones2015exploratory; @jones2016there.
+
 ![The partial dependence of several covariates (each considered separately) on the probability that a convict voted in the 2012 presidential election, given that they had registered to do so. Data is from @gerber2014can.](partial_dependence.png)
 
 Partial dependence, as described by @friedman2001greedy, estimates the marginal relationship between a subset of the covariates and the model's predictions by averaging over the marginal distribution of the compliment of this subset of the covariates. This approximation allows the display of the relationship between this subset of the covariates and the model's predictions even when there are many covariates which may interact. This functionality works with models fit by any of the aforementioned packages to any of the supported types of outcome variables. `partial_dependence` can be parallelized and also contains a number of additional parameters which allow the user to control this approximation. There is an associated plot function `plot_pd` which constructs plots for a wide variety of possible outputs from `partial_dependence` (e.g., when pairs of covariates are considered jointly, when each covariate is considered separately, when the outcome variable is categorical, etc).
