### 2016-10-13 #92 @zmjones
I added a doi for one article but couldn't find one for the Friedman 2001 article. I do have a jstor stable url that I could include though.

I also added two other papers which use this software, however they are unpublished. Is that ok? I see that it is necessary to talk about applications of the software to solve research problems.
