+++ paper/paper.bib
@@ -8,14 +8,19 @@
   publisher =	 {JSTOR}
 }
 
-@article{gerber2014can,
-  title =	 {Can Incarcerated Felons Be (Re) integrated into the
+@article {gerber2014can,
+  author =	 {Gerber, Alan S. and Huber, Gregory A. and Meredith,
+                  Marc and Biggers, Daniel R. and Hendry, David J.},
+  title =	 {Can Incarcerated Felons Be (Re)integrated into the
                   Political System? Results from a Field Experiment},
-  author =	 {Gerber, Alan S and Huber, Gregory A and Meredith,
-                  Marc and Biggers, Daniel R and Hendry, David J},
   journal =	 {American Journal of Political Science},
-  year =	 2014,
-  publisher =	 {Wiley Online Library}
+  volume =	 59,
+  number =	 4,
+  issn =	 {1540-5907},
+  url =		 {http://dx.doi.org/10.1111/ajps.12166},
+  doi =		 {10.1111/ajps.12166},
+  pages =	 {912--926},
+  year =	 2015,
 }
 
 @inproceedings{jones2015exploratory,
