### 2016-10-14 #92 @cMadan
It looks like the doi for Friedman (2001) is 10.1214/aos/1013203451 (see http://projecteuclid.org/euclid.aos/1013203451#info). The publisher is also incorrect in being listed as JSTOR.

I think it's fine to cite 'jones2015exploratory' as a conference paper, though I think the booktitle start with "Proceedings of" rather than "Prepared for". If you think there is a stable url, I would appreciate you including it here, but I'll leave that to your discretion.

For 'jones2016there', some additional details should be included, or it should be listed as @unpublished (rather than @article).
