### 2016-10-14 a8b8745
update citations
