+++ paper/paper.bib
@@ -1,11 +1,16 @@
 @article{friedman2001greedy,
-  title =	 {Greedy function approximation: a gradient boosting
-                  machine},
-  author =	 {Friedman, Jerome H},
-  journal =	 {Annals of statistics},
-  pages =	 {1189--1232},
-  year =	 2001,
-  publisher =	 {JSTOR}
+  author =	 "Friedman, Jerome H.",
+  doi =		 "10.1214/aos/1013203451",
+  journal =	 "The Annals of Statistics",
+  month =	 10,
+  number =	 5,
+  pages =	 "1189--1232",
+  publisher =	 "The Institute of Mathematical Statistics",
+  title =	 "Greedy function approximation: A gradient boosting
+                  machine.",
+  url =		 "http://dx.doi.org/10.1214/aos/1013203451",
+  volume =	 29,
+  year =	 2001
 }
 
 @article {gerber2014can,
@@ -25,13 +30,14 @@
 
 @inproceedings{jones2015exploratory,
   title =	 {Exploratory data analysis using random forests},
-  author =	 {Jones, Zachary and Linder, Fridolin},
-  booktitle =	 {Prepared for the 73rd annual MPSA conference},
-  year =	 2015
+  author =	 {Jones, Zachary M. and Linder, Fridolin},
+  booktitle =	 {Proceedings of the 73rd annual MPSA conference},
+  year =	 2015,
+  url = {http://zmjones.com/static/papers/rfss_manuscript.pdf}
 }
 
-@article{jones2016there,
+@unpublished{jones2016there,
   title =	 {Is There More Violence in the Middle?},
-  author =	 {Jones, Zachary M and Lupu, Yonatan},
+  author =	 {Jones, Zachary M. and Lupu, Yonatan},
   year =	 2016
 }
