### 2016-10-14 #92 @zmjones
Thanks. Made all of those changes.
