### 2016-10-14 e076621
update .ivar_points
