### 2016-10-18 #92 @cMadan
@PhilippPro - have you had a chance to take to try out the package since @zmjones has made the recent changes? Thanks again!
