### 2016-10-19 #92 @PhilippPro
I will finish the review by the end of the week as mentioned in the other issue. ;)
