### 2016-10-19 9774246
adds rf package citations
