+++ paper/paper.bib
@@ -33,11 +33,49 @@
   author =	 {Jones, Zachary M. and Linder, Fridolin},
   booktitle =	 {Proceedings of the 73rd annual MPSA conference},
   year =	 2015,
-  url = {http://zmjones.com/static/papers/rfss_manuscript.pdf}
+  url =
+                  {http://zmjones.com/static/papers/rfss_manuscript.pdf}
 }
 
 @unpublished{jones2016there,
   title =	 {Is There More Violence in the Middle?},
   author =	 {Jones, Zachary M. and Lupu, Yonatan},
   year =	 2016
+}
+
+@article{randomForest,
+  title =	 {Classification and Regression by randomForest},
+  author =	 {Andy Liaw and Matthew Wiener},
+  journal =	 {R News},
+  year =	 2002,
+  volume =	 2,
+  number =	 3,
+  pages =	 {18-22},
+  url =		 {http://CRAN.R-project.org/doc/Rnews/},
+}
+
+@article{randomForestSRC,
+  title =	 {Random forests for survival, regression and
+                  classification (RF-SRC)},
+  author =	 {Ishwaran, H and Kogalur, U},
+  journal =	 {URL
+                  http://cran.r-project.org/web/packages/randomForestSRC/. R
+                  package version},
+  volume =	 1,
+  year =	 2013
+}
+
+@article{party,
+  title =	 {Unbiased recursive partitioning: A conditional
+                  inference framework},
+  author =	 {Hothorn, Torsten and Hornik, Kurt and Zeileis,
+                  Achim},
+  journal =	 {Journal of Computational and Graphical statistics},
+  volume =	 15,
+  number =	 3,
+  pages =	 {651--674},
+  year =	 2006,
+  publisher =	 {Taylor \& Francis},
+  url =		 {http://dx.doi.org/10.1198/106186006X133933},
+  doi =		 {10.1198/106186006X133933}
 }
