### 2016-10-19 86c3196
removes unsupervised code/docs
