+++ pkg/R/unsup.R
@@ -1,34 +1,0 @@
-#' augment data by a univariate boostrap of the features
-#'
-#' @param data a matrix or data.frame
-#' @return a data.frame with row dimension twice that of data, with an additional factor column
-#' indicating whether the data is from the input, data (labelled "observed"), or was generated by
-#' the univariate bootstrap (labelled "synthetic")
-#' @export
-augment_data <- function(data) {
-  stopifnot((is.matrix(data) | is.data.frame(data)) & nrow(data) > 1)
-  augmented_data <- rbind(data, apply(data, 2, function(x) sample(x, length(x), TRUE)))
-  label <- as.factor(c(rep("observed", nrow(data)), rep("synthetic", nrow(data))))
-  as.data.frame(cbind(label, augmented_data))
-}
-#' compute the distance between features using a random forest
-#'
-#' @param data a matrix or data.frame for training
-#' @param newdata a matrix or data.frame to compute distance using the training data
-#' @param method the reandom forest implementation to use: "party", "randomForest", or "randomForestSRC":
-#' default is "party"
-#' @param ... additional arguments to pass to the random forest method used for training
-#' @examples
-#'
-#' data(iris)
-#' d <- randomforest_dist(iris[, -ncol(iris)], method = "randomForest")
-#' @export
-randomforest_dist <- function(data, newdata = NULL, method = "party", ...) {
-  n <- if (is.null(newdata)) nrow(data) else nrow(data)
-
-  fit <- switch(method,
-                "party" = party::cforest(label ~ ., data = augment_data(data), ...),
-                "randomForest" = randomForest::randomForest(label ~ ., augment_data(data), proximity = TRUE, ...),
-                "randomForestSRC" = randomForestSRC::rfsrc(label ~ ., augment_data(data), proximity = TRUE, ...))
-  extract_proximity(fit, newdata = newdata)[1:n, 1:n]
-}
