+++ pkg/NAMESPACE
@@ -9,14 +9,12 @@ S3method(partial_dependence,rfsrc)
 S3method(variable_importance,RandomForest)
 S3method(variable_importance,randomForest)
 S3method(variable_importance,rfsrc)
-export(augment_data)
 export(extract_proximity)
 export(partial_dependence)
 export(plot_imp)
 export(plot_pd)
 export(plot_pred)
 export(plot_prox)
-export(randomforest_dist)
 export(variable_importance)
 import(assertthat)
 import(ggplot2)
