### 2016-10-19 c6c60e0
error from earlier commit, unclear why travis didn't catch this. examples fail for pd.R fixes issue #50
