### 2016-10-19 6757330
fixes proximity extraction/plot example
