### 2016-10-19 9782cef
fixed travis config
