+++ .travis.yml
@@ -2,10 +2,16 @@ language: r
 sudo: required
 cache: packages
 
-install:
-  - Rscript -e 'install.packages(c("devtools", "roxygen2", "testthat", "knitr", "rmarkdown")); devtools::install_deps("pkg")'
-script:
-  - Rscript -e 'devtools::check("pkg")'
+before_install:
+  - cd pkg
+
+r_packages:
+  - testthat
+  - rmarkdown
+  - knitr
+  - party
+  - randomForest
+  - randomForestSRC
 
 notifications:
   email:
