### 2016-10-20 95597d4
add some references, clean up the readme some more
