+++ pkg/man/plot_imp.Rd
@@ -40,7 +40,7 @@ imp <- variable_importance(fit, var = colnames(iris)[-5],
 plot_imp(imp)
 
 imp <- variable_importance(fit, var = colnames(iris)[-5], type = "aggregate",
-                           interaction = FALSE, nperm = 2, oob = TRUE, data = iris)
+                           interaction = FALSE, nperm = 2, data = iris)
 plot_imp(imp)
 
 data(swiss)
