### 2016-10-20 #92 @cMadan
@PhilippPro - awesome, thank you!

@zmjones - everything looks good to me! I think the only thing left then is to archive the repository in a service such as Zenodo or figshare and let us know the associated doi.

@arfon - just pinging you to let you know that we're almost all set here!
