### 2016-10-20 #92 @zmjones
done. i added a link to the zenodo doi in the readme too

 DOI: 10.5281/zenodo.162238
