### 2016-10-21 #92 @arfon
@zmjones - it looks like there aren't any ORCID fields present in the `paper.md` metadata. We like authors to always have these if they have an academic affiliation - can you add these please?
