+++ paper/paper.md
@@ -7,6 +7,7 @@ tags:
 authors:
  - name: Zachary M. Jones
    affiliation: 1
+   orcid: 0000-0002-7523-0471
  - name: Fridolin J. Linder
    affiliation: 1
 affiliations:
