### 2016-10-21 #92 @arfon
:zap: thanks. Just like this:

```  authors:
   - name: Arfon M Smith
     orcid: 0000-0002-3957-2474
     affiliation: 1
   - name: Mickey Mouse
     orcid: 0000-0000-0000-1234
     affiliation: 2
```
