### 2016-10-21 be0bedd
add joss paper status
