### 2016-10-23 #92 @arfon
@zmjones - your paper is now accepted into JOSS and the DOI is http://dx.doi.org/10.21105/joss.00092

@PhilippPro many thanks for reviewing this one!
